// RydrDriver/src/API.ts
export type ModelStringInput = {
  ne?: string | null;
  eq?: string | null;
  le?: string | null;
  lt?: string | null;
  ge?: string | null;
  gt?: string | null;
  contains?: string | null;
  notContains?: string | null;
  between?: Array<string | null> | null;
  beginsWith?: string | null;
  attributeExists?: boolean | null;
  attributeType?: ModelAttributeTypes | null;
  size?: ModelSizeInput | null;
};
// ...
export type ModelFloatInput = {
  ne?: number | null;
  eq?: number | null;
  le?: number | null;
  lt?: number | null;
  ge?: number | null;
  gt?: number | null;
  between?: Array<number | null> | null;
  attributeExists?: boolean | null;
  attributeType?: ModelAttributeTypes | null;
};

export type User = {
  __typename: 'User';
  id: string;
  username: string;
  given_name: string;
  family_name: string;
  phone_number: string;
  currentLat: number;
  currentLng: number;
  heading: number;
  email: string;
  orders?: ModelOrderConnection | null;
  createdAt: string;
  updatedAt: string;
};

export type ModelOrderConnection = {
  __typename: 'ModelOrderConnection';
  items: Array<Order | null>;
  nextToken?: string | null;
};

export type Order = {
  __typename: 'Order';
  id: string;
  createdAt: string;
  type: string;
  status: string;
  origin: string;
  destination: string;
  originAddress: string;
  destinationAddress: string;
  fare: number;
  scheduleDate: string;
  passengerNumber: string;
  paymentType: string;
  notes: string;
  designate: string;
  userId: string;
  user?: User | null;
  driverId?: string | null;
  driver?: Driver | null;
  updatedAt: string;
};

export type Driver = {
  __typename: 'Driver';
  id: string;
  username: string;
  given_name: string;
  family_name: string;
  phone_number: string;
  currentLat: number;
  currentLng: number;
  heading: number;
  email: string;
  type: string;
  isActive?: boolean | null;
  orders?: ModelOrderConnection | null;
  createdAt: string;
  updatedAt: string;
};
// ...
export type ModelBooleanInput = {
  ne?: boolean | null;
  eq?: boolean | null;
  attributeExists?: boolean | null;
  attributeType?: ModelAttributeTypes | null;
};
// ...
export type ModelIDInput = {
  ne?: string | null;
  eq?: string | null;
  le?: string | null;
  lt?: string | null;
  ge?: string | null;
  gt?: string | null;
  contains?: string | null;
  notContains?: string | null;
  between?: Array<string | null> | null;
  beginsWith?: string | null;
  attributeExists?: boolean | null;
  attributeType?: ModelAttributeTypes | null;
  size?: ModelSizeInput | null;
};
// ...
export type ModelUserFilterInput = {
  id?: ModelIDInput | null;
  username?: ModelStringInput | null;
  given_name?: ModelStringInput | null;
  family_name?: ModelStringInput | null;
  phone_number?: ModelStringInput | null;
  currentLat?: ModelFloatInput | null;
  currentLng?: ModelFloatInput | null;
  heading?: ModelFloatInput | null;
  email?: ModelStringInput | null;
  and?: Array<ModelUserFilterInput | null> | null;
  or?: Array<ModelUserFilterInput | null> | null;
  not?: ModelUserFilterInput | null;
};

export type ModelUserConnection = {
  __typename: 'ModelUserConnection';
  items: Array<User | null>;
  nextToken?: string | null;
};

export type ModelDriverFilterInput = {
  id?: ModelIDInput | null;
  username?: ModelStringInput | null;
  given_name?: ModelStringInput | null;
  family_name?: ModelStringInput | null;
  phone_number?: ModelStringInput | null;
  currentLat?: ModelFloatInput | null;
  currentLng?: ModelFloatInput | null;
  heading?: ModelFloatInput | null;
  email?: ModelStringInput | null;
  type?: ModelStringInput | null;
  isActive?: ModelBooleanInput | null;
  and?: Array<ModelDriverFilterInput | null> | null;
  or?: Array<ModelDriverFilterInput | null> | null;
  not?: ModelDriverFilterInput | null;
};

export type ModelDriverConnection = {
  __typename: 'ModelDriverConnection';
  items: Array<Driver | null>;
  nextToken?: string | null;
};

export type ModelOrderFilterInput = {
  id?: ModelIDInput | null;
  createdAt?: ModelStringInput | null;
  type?: ModelStringInput | null;
  status?: ModelStringInput | null;
  origin?: ModelStringInput | null;
  destination?: ModelStringInput | null;
  originAddress?: ModelStringInput | null;
  destinationAddress?: ModelStringInput | null;
  fare?: ModelFloatInput | null;
  scheduleDate?: ModelStringInput | null;
  passengerNumber?: ModelStringInput | null;
  paymentType?: ModelStringInput | null;
  notes?: ModelStringInput | null;
  designate?: ModelStringInput | null;
  userId?: ModelIDInput | null;
  driverId?: ModelIDInput | null;
  and?: Array<ModelOrderFilterInput | null> | null;
  or?: Array<ModelOrderFilterInput | null> | null;
  not?: ModelOrderFilterInput | null;
};

export enum ModelSortDirection {
  ASC = 'ASC',
  DESC = 'DESC',
}
// ...
export type GetUserQueryVariables = {
  id: string;
};

export type GetUserQuery = {
  getUser?: {
    __typename: 'User';
    id: string;
    username: string;
    given_name: string;
    family_name: string;
    phone_number: string;
    currentLat: number;
    currentLng: number;
    heading: number;
    email: string;
    orders?: {
      __typename: 'ModelOrderConnection';
      nextToken?: string | null;
    } | null;
    createdAt: string;
    updatedAt: string;
  } | null;
};

export type ListUsersQueryVariables = {
  filter?: ModelUserFilterInput | null;
  limit?: number | null;
  nextToken?: string | null;
};

export type ListUsersQuery = {
  listUsers?: {
    __typename: 'ModelUserConnection';
    items: Array<{
      __typename: 'User';
      id: string;
      username: string;
      given_name: string;
      family_name: string;
      phone_number: string;
      currentLat: number;
      currentLng: number;
      heading: number;
      email: string;
      createdAt: string;
      updatedAt: string;
    } | null>;
    nextToken?: string | null;
  } | null;
};

export type GetDriverQueryVariables = {
  id: string;
};

export type GetDriverQuery = {
  getDriver?: {
    __typename: 'Driver';
    id: string;
    username: string;
    given_name: string;
    family_name: string;
    phone_number: string;
    currentLat: number;
    currentLng: number;
    heading: number;
    email: string;
    type: string;
    isActive?: boolean | null;
    orders?: {
      __typename: 'ModelOrderConnection';
      nextToken?: string | null;
    } | null;
    createdAt: string;
    updatedAt: string;
  } | null;
};

export type ListDriversQueryVariables = {
  filter?: ModelDriverFilterInput | null;
  limit?: number | null;
  nextToken?: string | null;
};

export type ListDriversQuery = {
  listDrivers?: {
    __typename: 'ModelDriverConnection';
    items: Array<{
      __typename: 'Driver';
      id: string;
      username: string;
      given_name: string;
      family_name: string;
      phone_number: string;
      currentLat: number;
      currentLng: number;
      heading: number;
      email: string;
      type: string;
      isActive?: boolean | null;
      createdAt: string;
      updatedAt: string;
    } | null>;
    nextToken?: string | null;
  } | null;
};

export type GetOrderQueryVariables = {
  id: string;
};

export type GetOrderQuery = {
  getOrder?: {
    __typename: 'Order';
    id: string;
    createdAt: string;
    type: string;
    status: string;
    origin: string;
    destination: string;
    originAddress: string;
    destinationAddress: string;
    fare: number;
    scheduleDate: string;
    passengerNumber: string;
    paymentType: string;
    notes: string;
    designate: string;
    userId: string;
    user?: {
      __typename: 'User';
      id: string;
      username: string;
      given_name: string;
      family_name: string;
      phone_number: string;
      currentLat: number;
      currentLng: number;
      heading: number;
      email: string;
      createdAt: string;
      updatedAt: string;
    } | null;
    driverId?: string | null;
    driver?: {
      __typename: 'Driver';
      id: string;
      username: string;
      given_name: string;
      family_name: string;
      phone_number: string;
      currentLat: number;
      currentLng: number;
      heading: number;
      email: string;
      type: string;
      isActive?: boolean | null;
      createdAt: string;
      updatedAt: string;
    } | null;
    updatedAt: string;
  } | null;
};

export type ListOrdersQueryVariables = {
  filter?: ModelOrderFilterInput | null;
  limit?: number | null;
  nextToken?: string | null;
};

export type ListOrdersQuery = {
  listOrders?: {
    __typename: 'ModelOrderConnection';
    items: Array<{
      __typename: 'Order';
      id: string;
      createdAt: string;
      type: string;
      status: string;
      origin: string;
      destination: string;
      originAddress: string;
      destinationAddress: string;
      fare: number;
      scheduleDate: string;
      passengerNumber: string;
      paymentType: string;
      notes: string;
      designate: string;
      userId: string;
      driverId?: string | null;
      updatedAt: string;
    } | null>;
    nextToken?: string | null;
  } | null;
};

export type OrdersByUserIdQueryVariables = {
  userId: string;
  sortDirection?: ModelSortDirection | null;
  filter?: ModelOrderFilterInput | null;
  limit?: number | null;
  nextToken?: string | null;
};

export type OrdersByUserIdQuery = {
  ordersByUserId?: {
    __typename: 'ModelOrderConnection';
    items: Array<{
      __typename: 'Order';
      id: string;
      createdAt: string;
      type: string;
      status: string;
      origin: string;
      destination: string;
      originAddress: string;
      destinationAddress: string;
      fare: number;
      scheduleDate: string;
      passengerNumber: string;
      paymentType: string;
      notes: string;
      designate: string;
      userId: string;
      driverId?: string | null;
      updatedAt: string;
    } | null>;
    nextToken?: string | null;
  } | null;
};

export type OrdersByDriverIdQueryVariables = {
  driverId: string;
  sortDirection?: ModelSortDirection | null;
  filter?: ModelOrderFilterInput | null;
  limit?: number | null;
  nextToken?: string | null;
};

export type OrdersByDriverIdQuery = {
  ordersByDriverId?: {
    __typename: 'ModelOrderConnection';
    items: Array<{
      __typename: 'Order';
      id: string;
      createdAt: string;
      type: string;
      status: string;
      origin: string;
      destination: string;
      originAddress: string;
      destinationAddress: string;
      fare: number;
      scheduleDate: string;
      passengerNumber: string;
      paymentType: string;
      notes: string;
      designate: string;
      userId: string;
      driverId?: string | null;
      updatedAt: string;
    } | null>;
    nextToken?: string | null;
  } | null;
};